pyproject.toml
[tool.poetry]
authors = ["Gustav Odinger <65498475+gustavwilliam@users.noreply.github.com>"]
description = ""
name = "tuif"
version = "0.1.0"

[tool.poetry.dependencies]
PyYAML = "^5.4.1"
python = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.6b0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
